+++ src/main/java/com/pro/finance/selffinanceapp/repository/IncomeRepository.java
@@ -6,6 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface IncomeRepository extends JpaRepository<Income, Long> {
@@ -15,5 +16,5 @@ public interface IncomeRepository extends JpaRepository<Income, Long> {
 
     // 🔹 Get total income (used in FinancialHealthService)
     @Query("SELECT COALESCE(SUM(i.amount), 0) FROM Income i WHERE i.user.id = :userId")
-    double getTotalIncome(@Param("userId") Long userId);
+    BigDecimal getTotalIncome(@Param("userId") Long userId);
 }
